FEAT : mockController 삭제 및 도메인별 Controller.class 추가
